+++ src/app.jsx
@@ -1,16 +1,14 @@
-import { Header, Footer, Layout } from "@/components/common";
+import { Layout } from "@/components/common";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import LandingPage from "./pages/LandingPage";
+import { Landing } from "@/pages";
 
 function App() {
   return (
     <Router>
       <Layout>
-        <Header />
         <Routes>
-          <Route path="/" element={<LandingPage />} />
+          <Route path="/" element={<Landing />} />
         </Routes>
-        <Footer />
       </Layout>
     </Router>
   );
