+++ src/components/common/index.jsx
@@ -1,6 +1,3 @@
-export { default as Button } from "./button";
-export { default as Footer } from "./footer";
-export { default as Hero } from "./hero";
-export { default as About } from "./about";
-export { default as Header } from "./header";
 export { default as Layout } from "./layout";
+
+export * from "./buttons";
