--- src/components/landing.jsx
@@ -1,5 +1,0 @@
-const Landing = () => {
-  return <div></div>;
-};
-
-export default Landing;
