+++ src/components/landing/index.jsx
@@ -1,0 +1,3 @@
+export { default as Console } from "./console";
+export { default as Hero } from "./hero";
+export { default as Vision } from "./vision";
