+++ src/components/landing/vision/data.json
@@ -1,0 +1,30 @@
+[
+  {
+    "id": 1,
+    "imgType": "npm",
+    "details": "@sliit-foss/functions",
+    "date": "May 2023",
+    "logoFirst": true
+  },
+  {
+    "id": 2,
+    "imgType": "npm",
+    "details": "@sliit-foss/babel-plugin-transform-trace",
+    "date": "June 2023",
+    "logoFirst": false
+  },
+  {
+    "id": 3,
+    "imgType": "npm",
+    "details": "@sliit-foss/timekeeper",
+    "date": "August 2023",
+    "logoFirst": true
+  },
+  {
+    "id": 4,
+    "imgType": "vscode",
+    "details": "Timekeeper extension",
+    "date": "December 2023",
+    "logoFirst": false
+  }
+]
