--- src/pages/LandingPage.jsx
@@ -1,8 +1,0 @@
-import OurVision from "../components/ourVision";
-import timelineData from "../data/timelineData.json";
-
-const LandingPage = () => {
-  return <OurVision dataList={timelineData.timeline} />;
-};
-
-export default LandingPage;
