+++ src/pages/landing.jsx
@@ -1,0 +1,13 @@
+import { Hero, Console, Vision } from "@/components/landing";
+
+const Landing = () => {
+  return (
+    <>
+      <Hero />;
+      <Console />;
+      <Vision />;
+    </>
+  );
+};
+
+export default Landing;
