+++ src/styles/index.css
@@ -19,13 +19,3 @@
     scroll-behavior: smooth;
   }
 }
-
-@layer components {
-  .gradient-text {
-    @apply bg-gradient-to-r from-gl to-gr bg-clip-text text-transparent;
-  }
-
-  .btn-primary {
-    @apply flex items-center font-semibold text-[24px] font-sf space-x-2 mx-auto border px-4 py-1 rounded-2xl border-black border-opacity-10;
-  }
-}
